conduct unit test for system, system module and permission fix models and fix bug in authorization middleware

+++ app/Http/Controllers/UmisAndEmployeeManagement/SystemModuleController.php
@@ -2,12 +2,18 @@
 
 namespace App\Http\Controllers\UmisAndEmployeeManagement;
 
+use App\Http\Controllers\Controller;
+
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Log;
 use App\Services\RequestLogger;
+use App\Http\Resources\ModulePermissionResource;
 use App\Http\Resources\SystemModuleResource;
 use App\Http\Requests\SystemModuleRequest;
+use App\Models\ModulePermission;
+use App\Models\Permission;
+use App\Models\System;
 use App\Models\SystemModule;
 use App\Models\SystemLogs;
 
@@ -24,12 +30,12 @@ class SystemModuleController extends Controller
         $this->requestLogger = $requestLogger;
     }
 
-    public function index(Request $request)
+    public function index($id, Request $request)
     {
         try{
-            $system_modules = SystemModule::all();
+            $system_modules = SystemModule::where('system_id', $id)->get();
 
-            $this->registerSystemLogs($request, null, true, 'Success in fetching '.$this->PLURAL_MODULE_NAME.'.');
+            $this->registerSystemLogs($request, $id, true, 'Success in system module fetching '.$this->PLURAL_MODULE_NAME.'.');
 
             return response()->json([
                 'data' => SystemModuleResource::collection($system_modules),
@@ -41,17 +47,21 @@ class SystemModuleController extends Controller
         }
     }
     
-    public function store(SystemModuleRequest $request)
+    public function store($id, SystemModuleRequest $request)
     {
         try{
+            $system = System::find($id);
+
+            if(!$system)
+            {
+                return response()->json(['messsage' => 'No system found for id '.$id. ' .'], Response::HTTP_NOT_FOUND);
+            }
+
             $cleanData = [];
+            $cleanData['system_id'] = $system->id;
 
             foreach ($request->all() as $key => $value) {
-                if (is_bool($value) || $value === null) {
-                    $cleanData[$key] = $value;
-                } else {
-                    $cleanData[$key] = strip_tags($value);
-                }
+                $cleanData[$key] = strip_tags($value);
             }
 
             $system_module = SystemModule::create($cleanData);
@@ -83,6 +93,15 @@ class SystemModuleController extends Controller
     public function addPermission($id, Request $request)
     {
         try{
+            $new_module_permission = [];
+
+            $system_module = SystemModule::find($id);
+            
+            if(!$system_module)
+            {
+                return response()->json(['message' => 'No record found for system module id '.$id.' .'], Response::HTTP_NOT_FOUND);
+            }
+
             $permissions = $request->input('permissions');
             
             $failed = [];
@@ -91,13 +110,24 @@ class SystemModuleController extends Controller
                 $module_permission = ModulePermission::where('system_module_id',$id)->where('permission_id', $value)->first();
 
                 try{
+                    $permission = Permission::find($value);
+
+                    if(!$permission)
+                    {
+                        $fail_registration = [
+                            'permission_id' => $value,
+                            'remarks' => 'No record found for permission with id '.' .'
+                        ];
+
+                        $failed[] = $fail_registration;
+                        continue;
+                    }
+
                     if(!$module_permission){
-                        $system_module = SystemModule::find($id);
-                        $permission = Permission::find($value);
     
                         $code = $system_module['code'].' '.$permission['action'];
     
-                        ModulePermission::create([
+                        $new_module_permission[] =  ModulePermission::create([
                             'system_module_id' => $system_module['id'],
                             'permission_id' => $permission['id'],
                             'code' => $code
@@ -122,17 +152,27 @@ class SystemModuleController extends Controller
                 }
             }
 
+            if(count($failed) === count($permissions))
+            {
+                $this->registerSystemLogs($request, $id, false, 'Failed in creating module permission '.$this->SINGULAR_MODULE_NAME.'.');
+
+                return response()->json(['message' => "Failed to register all permissions for this module id ".$id." ."], Response::HTTP_OK);
+            }
+
             if(count($failed) > 0)
             {
                 $this->registerSystemLogs($request, $id, true, 'Success in creating module permission but some failed '.$this->SINGULAR_MODULE_NAME.'.');
 
-                return response()->json(['data' => $failed, 'message' => "Some permission did not register."], Response::HTTP_OK);
+                return response()->json(['data' => ModulePermissionResource::collection($new_module_permission) , 'failed'  => $failed, 'message' => "Some permission did not register."], Response::HTTP_OK);
             }
 
 
             $this->registerSystemLogs($request, $id, true, 'Success in creating module permission '.$this->SINGULAR_MODULE_NAME.'.');
 
-            return response()->json(['message' => 'New permission added to system module.'], Response::HTTP_OK);
+            return response()->json([
+                'data' => ModulePermissionResource::collection($new_module_permission),
+                'message' => 'New permission added to system module.'
+            ], Response::HTTP_OK);
         }catch(\Throwable $th){
             $this->requestLogger->errorLog($this->CONTROLLER_NAME,'addPermission', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
@@ -142,7 +182,7 @@ class SystemModuleController extends Controller
     public function show($id, Request $request)
     {
         try{
-            $system_module = SystemModule::findOrFail($id);
+            $system_module = SystemModule::find($id);
 
             if(!$system_module)
             {
@@ -156,7 +196,7 @@ class SystemModuleController extends Controller
                 'message' => 'System module record retrieved.'
             ], Response::HTTP_OK);
         }catch(\Throwable $th){
-            $this->requestLogger->errorLog($this->CONTROLLER_NAME,'index', $th->getMessage());
+            $this->requestLogger->errorLog($this->CONTROLLER_NAME,'show', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -174,11 +214,7 @@ class SystemModuleController extends Controller
             $cleanData = [];
 
             foreach ($request->all() as $key => $value) {
-                if (is_bool($value) || $value === null) {
-                    $cleanData[$key] = $value;
-                } else {
-                    $cleanData[$key] = strip_tags($value);
-                }
+                $cleanData[$key] = strip_tags($value);
             }
 
             $system_module->update($cleanData);
@@ -190,7 +226,7 @@ class SystemModuleController extends Controller
                 'message' => 'System module record updated'
             ], Response::HTTP_OK);
         }catch(\Throwable $th){
-            $this->requestLogger->errorLog($this->CONTROLLER_NAME,'index', $th->getMessage());
+            $this->requestLogger->errorLog($this->CONTROLLER_NAME,'update', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -205,13 +241,20 @@ class SystemModuleController extends Controller
                 return response()->json(['message' => 'No record found.'], Response::HTTP_NOT_FOUND);
             }
 
+            $module_permissions = $system_module->modulePermissions;
+
+            if(count($module_permissions)>0){    
+                $this->registerSystemLogs($request, $id, true, 'Failed in deleting cause this data in use with by other record '.$this->SINGULAR_MODULE_NAME.'.');
+                return response()->json(['message' => "There are data using this system module, can't, delete this record."], Response::HTTP_OK); 
+            }
+
             $system_module->delete();
             
             $this->registerSystemLogs($request, $id, true, 'Success in deleting '.$this->SINGULAR_MODULE_NAME.'.');
             
             return response()->json(['message' => 'System module record deleted.'], Response::HTTP_OK);
         }catch(\Throwable $th){
-            $this->requestLogger->errorLog($this->CONTROLLER_NAME,'index', $th->getMessage());
+            $this->requestLogger->errorLog($this->CONTROLLER_NAME,'destroy', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -221,7 +264,7 @@ class SystemModuleController extends Controller
         $ip = $request->ip();
         $user = $request->user;
         $permission = $request->permission;
-        list($action, $module) = explode(' ', $permission);
+        list($module, $action) = explode(' ', $permission);
 
         SystemLogs::create([
             'employee_profile_id' => $user->id,
@@ -230,7 +273,7 @@ class SystemModuleController extends Controller
             'module' => $module,
             'status' => $status,
             'remarks' => $remarks,
-            'ip_system_module' => $ip
+            'ip_address' => $ip
         ]);
     }
 }
